# doc/api.yaml
openapi: 3.0.3
info:
  description: |-
    Each user will be presented with a stream of photos in reverse chronological order, with information
    about when it was uploaded (date and time) and how many likes and comments it has. Users can place
    (and later remove) a “like” to photos from other users. Also, users can add comments to any photo
    (even those uploaded by themself). Comments can be removed by the author only.
    Users can ban other users. If Alice (user) bans Eve (user), Eve won’t be able to see any information
    about Alice. Alice can decide to remove the ban at any moment.
    Users will have their own personal profiles. The personal profile page for user detail: user’s photos (in
    reverse chronological order), their followers and following, and how many photos have been uploaded.
    Each user can change his/her own username, upload photos, remove photos, and follow/unfollow
    other users. Removal of a photo will also remove likes and comments.
    Users profiles can be searched via username.
    The login is perfomed by specifying the username.
  title: WASAPhoto API
  version: 1.0.0
  contact:
    name: Lucian Dorin Crainic
    email: crainic.1938430@studenti.uniroma1.it
    url: https://github.com/LucianCrainic
  license:
    name: MIT
...
paths:
...
  /users/{username}/ban/{banId}:
    parameters:
      - $ref: "#/components/parameters/username"
      - $ref: "#/components/parameters/banID"
    #* banUser
    put: 
      security:
      - bearerAuth : []
      tags: ["Ban"]
      summary: Ban user
      description: |-
       Ban an user.
      operationId: banUser
      responses:
        "201": 
          description: User banned successfully.
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/ban"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    #* unbanUser
    delete: 
      security:
      - bearerAuth : []
      tags: ["Ban"]
      summary: Unban user
      description: |-
        Unban an user.
      operationId: unbanUser

      responses:
       "204": { description: Ban deleted successfully. }
       "400": { $ref: "#/components/responses/BadRequest" }
       "401": { $ref: "#/components/responses/UnauthorizedError" }
       "500": { $ref: "#/components/responses/InternalServerError" }
  /users/{username}/ban:
    parameters:
      - $ref: "#/components/parameters/username"
    #* getBans
    get:
      security:
      - bearerAuth : []
      tags: ["Ban"]
      summary: Return ban
      operationId: getBans
      description: |-
        Return an user ban.
      responses:
        "200": 
          description: User bans retrived successfully.
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/ban"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /users/{username}/follow/{followId}:
    parameters:
      - $ref: "#/components/parameters/username"
      - $ref: "#/components/parameters/followID"
    #* followUser
    put: 
      security:
      - bearerAuth : []
      tags: ["Follow"]
      summary: Follow user
      description: |-
        Folow an user.
      operationId: followUser
      responses:
        "201": 
          description: User followed successfully.
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/follow"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    #* unfollowUser
    delete: 
      security:
      - bearerAuth : []
      tags: ["Follow"]
      summary: Unfollow user
      description: |-
       Unfollow an user.
      operationId: unfollowUser
      responses:
       "204": { description: Follow deleted successfully. }
       "400": { $ref: "#/components/responses/BadRequest" }
       "401": { $ref: "#/components/responses/UnauthorizedError" }
       "500": { $ref: "#/components/responses/InternalServerError" }
  /users/{username}/follow:
    parameters:
      - $ref: "#/components/parameters/username"
    #* getFollowers
    get:
      security:
      - bearerAuth : []
      tags: ["Follow"]
      summary: Return follow
      operationId: getFollowers
      description: |-
        Return an user follow.
      responses:
        "200": 
          description: User followers retrived successfully.
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/follow"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
  /users/{username}/photo/{photoId}/like/{likeId}:
    parameters:
      - $ref: "#/components/parameters/username"
      - $ref : "#/components/parameters/photoID"
      - $ref : "#/components/parameters/likeID"
    #* likePhoto
    put: 
      security:
      - bearerAuth : []
      tags: ["Like"]
      operationId: likePhoto
      summary: Like photo
      description: |-
        Add a like to a specific photo.
      responses:
        "201":
          description: Like added successfully.
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/like"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    #* unlikePhoto
    delete: 
      security:
      - bearerAuth : []
      tags: ["Like"]
      operationId: unlikePhoto
      summary: Delete like
      description: |-
        Remove a like from a specific photo.
      responses:
        "204": { description: Like deleted successfully. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /users/{username}/photo/{photoId}/like:
    parameters:
      - $ref: "#/components/parameters/username"
      - $ref : "#/components/parameters/photoID"
    #* getLike
    get:
      security:
      - bearerAuth : []
      tags: ["Like"]
      summary: Return like
      description: |-
        Return a like from a specific photo.
      operationId: getLikes
      responses:
        "200": 
          description: Photo like retrived successfully.
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/like"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /users/{username}/photo/{photoId}/comment/{commentId}:
    parameters:
      - $ref: "#/components/parameters/username"
      - $ref: "#/components/parameters/photoID"
      - $ref: "#/components/parameters/commentID"
    #* commentPhoto
    put:
      security:
      - bearerAuth : []
      tags: ["Comment"]
      summary: Add comment
      operationId: commentPhoto
      description: |-
       Add a comment under a specific photo.
      requestBody:
        description: Comment 
        content: 
          application/json:
            schema:
              $ref: "#/components/schemas/comment"
        required: true
      responses:
        "201":
          description:  Comment added successfully.
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/commentsList"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    #* uncommentPhoto
    delete:
     security:
     - bearerAuth : []
     tags: ["Comment"]
     operationId: uncommentPhoto
     summary: Remove comment
     description: |-
       Deletete a comment under a specific photo.
     responses:
       "204": { description: Comment deleted successfully. }
       "400": { $ref: "#/components/responses/BadRequest" }
       "401": { $ref: "#/components/responses/UnauthorizedError" }
       "500": { $ref: "#/components/responses/InternalServerError" }
  /users/{username}/photo/{photoId}/comment:
    parameters:
      - $ref: "#/components/parameters/username"
      - $ref: "#/components/parameters/photoID"
    #* getComments
    get:
      security:
      - bearerAuth : []
      tags: ["Comment"]
      summary: Return comments
      description: |-
        Return all comments under a specific photo.
      operationId: getComments
      responses:
        "200": 
          description: Photo comments retrived successfully.
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/commentsList"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" } 
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses: 
    UnauthorizedError:
      description: Access token is missing or invalid
    BadRequest:
      description: The request was not compliant with the documentation (eg. missing fields, etc).
    InternalServerError:
      description: The server encountered an internal error. Further info in server logs.
  schemas:
...
    ban:
      title: Ban 
      description: Ban schema.
      type: object
      properties:
        banId:
          description: banID schema.
          type: string
          pattern: '^.*?$'
          example: "#B1234"
          minLength: 3
          maxLength: 16
        bannedId:
          description: Unique user identifier.
          type: string
          pattern: '^.*?$'
          example: "#U1234"
          minLength: 3
          maxLength: 16
        userId:
          description: Unique user identifier.
          type: string
          pattern: '^.*?$'
          example: "#U1234"
          minLength: 3
          maxLength: 16
    follow:
      title: Follow 
      description: Follow schema.
      type: object
      properties:
        followId:
          description: followID schema.
          type: string
          pattern: '^.*?$'
          example: "#F1234"
          minLength: 3
          maxLength: 16
        followedId:
          description: Unique user identifier.
          type: string
          pattern: '^.*?$'
          example: "#U1234"
          minLength: 3
          maxLength: 16
        userId:
          description: Unique user identifier.
          type: string
          pattern: '^.*?$'
          example: "#U1234"
          minLength: 3
          maxLength: 16
    commentsList: 
      title: Comments
      description: Comments schema.
      type: object
      properties:
        requestId:
          description: Unique user identifier.
          type: string
          pattern: '^.*?$'
          example: "#U1234"
          minLength: 3
          maxLength: 16
        photoId: 
          description:  Unique photo identifier.
          type: string
          pattern: '^.*?$'
          example: "#P1234"
          minLength: 3
          maxLength: 16
        photoOwner:
          description: Unique user identifier.
          type: string
          pattern: '^.*?$'
          example: "#U1234"
          minLength: 3
          maxLength: 16
        comments: 
          description: list of objects (each object is a comment).
          type: array
          minItems: 0
          maxItems: 99999
          items:
            type: object
            properties:
              id:
                description: Unique comment identifier.
                type: string
                pattern: '^.*?$'
                example: "#C1234"
                minLength: 3
                maxLength: 16
              userId:
                description: Unique user identifier.
                type: string
                pattern: '^.*?$'
                example: "#U5678" 
                minLength: 3
                maxLength: 16
              photoId: 
                description:  Unique photo identifier.
                type: string
                pattern: '^.*?$'
                example: "#P1234"
                minLength: 3
                maxLength: 16
              photoOwner:
                description: Unique user identifier.
                type: string
                pattern: '^.*?$'
                example: "#U1234"
                minLength: 3
                maxLength: 16
              photoOwnerUsername:
                description: User username.
                type: string
                pattern: '^.*?$'
                example: LucianCrainic
                minLength: 3
                maxLength: 16
              username:
                description: User username.
                type: string
                pattern: '^.*?$'
                example: LucianCrainic
                minLength: 3
                maxLength: 16
              content:
                description: Content of the comment.
                type: string
                pattern: '^.*?$'
                example: "Hello World!"
                minLength: 2
                maxLength: 30
    like:
      title: Like 
      description: Like schema.
      type: object
      properties:
        likeId:
          description: likeID schema.
          type: string
          pattern: '^.*?$'
          example: "#L1234"
          minLength: 3
          maxLength: 16
        userId:
          description: Unique user identifier.
          type: string
          pattern: '^.*?$'
          example: "#U5678" 
          minLength: 3
          maxLength: 16
        photoId: 
          description:  Unique photo identifier.
          type: string
          pattern: '^.*?$'
          example: "#P1234"
          minLength: 3
          maxLength: 16
        photoOwner:
          description: Unique user identifier.
          type: string
          pattern: '^.*?$'
          example: "#U1234"
          minLength: 3
          maxLength: 16
...
    comment:
      title: Comment
      type: object
      description: comment schema.
      properties:
        content:
          description: content
          type: string
          example: Hello World!
          pattern: '^.*?$'
          minLength: 1
          maxLength: 40  
  parameters:
    username:
      schema:
          description: username schema.
          type: string
          pattern: '^.*?$'
          example: LucianCrainic
          minLength: 5
          maxLength: 30
      name: username
      in: path
      required: true
      description: "**Username** of a WASAPhoto user."
    photoID:
      schema:
        description: photoID schema.
        type: string
        pattern: '^.*?$'
        example: "#P1234"
        minLength: 3
        maxLength: 16
      name: photoId 
      in: path
      required: true
      description: "**photoID** of a specific photo on a user profile."
    commentID:
      schema:
        description: commentID schema.
        type: string
        pattern: '^.*?$'
        example: "#C1234"
        minLength: 3
        maxLength: 16
      name: commentId
      in: path
      required: true
      description: "**commentID** of a single comment under a photo."
    likeID:
      schema:
        description: likeID schema.
        type: string
        pattern: '^.*?$'
        example: "#L1234"
        minLength: 3
        maxLength: 16
      name: likeId
      in: path
      required: true
      description: "**likeID** of a single comment under a photo."
    followID:
      schema:
        description: followID schema.
        type: string
        pattern: '^.*?$'
        example: "#F1234"
        minLength: 3
        maxLength: 16
      name: followId
      in: path
      required: true
      description: "**followID** of a single comment under a photo."
    banID:
      schema:
        description: banID schema.
        type: string
        pattern: '^.*?$'
        example: "#B1234"
        minLength: 3
        maxLength: 16
      name: banId
      in: path
      required: true
      description: "**banID** of a single comment under a photo."